Move bundle plugin handling under trytls.bundles
